se agrego la validacion con class-validator, se probaron todos los endpoints y funcionan solo faltaria modificar la nomenclatura de las url

+++ src/book/application/deleteBookUseCase.ts
@@ -1,5 +1,7 @@
 
+import { validate } from "class-validator";
 import { IBookRepositorio } from "../domain/bookRepository";
+import { ValidatorId } from "../domain/validations/books";
 
 
 export class DeleteBookUseCase {
@@ -8,6 +10,11 @@ export class DeleteBookUseCase {
     async run(
         uuid: string,
     ):Promise<string | null> {
+        let post = new ValidatorId(uuid)
+        const validation = await validate(post)
+        if (validation.length > 0) {
+            throw new Error(JSON.stringify(validation));
+        }
         
         try {
             const newBook = await this.bookRepository.deleteBook(uuid,)
